backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.122.0"}
psycopg = {extras = ["binary"], version = "^3.2.13"}
youtube-transcript-api = {version = "^1.2.3", python = ">=3.12,<3.15"}
openai = "^2.8.1"
python-dotenv = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
